Prepared a new allocator for parsing
- Added warning flags in the Makefile.
- Applied corrections based on the warnings.
- Crafted a new allocator and refactored the old one. Also preparing
  the reimplementation of the memory management during parsing.
- Removed unused functions in "lexer.c".

Parsing causes crash when new chunks are allocated because the pointers may
point to old locations that are no longer accessible.

+++ src/source.c
@@ -52,7 +52,7 @@ Source* restrict source) {
 		1,
 		source->length,
 		source_file)
-	!= source->length) {
+	!= (size_t) source->length) {
 		printf("Cannot read the source at \"");
 		printf(path);
 		printf("\".\n");
@@ -70,7 +70,7 @@ ERROR:
 		return false;
 	}
 
-	source->content[source->length + 1] = '\0';
+	source->content[(size_t) source->length + 1] = '\0';
 	return true;
 }
